Moved test to Tests solution
