+++ graphQL/GraphQLStrawberryShake/GraphQLStrawberryShake.sln
@@ -7,9 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQLStrawberryShake", "G
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQLStrawberryShake.Server", "GraphQLStrawberryShake.Server\GraphQLStrawberryShake.Server.csproj", "{EEDE42E8-5BD9-4B16-91FE-8D68AF8B1343}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQLStrawberryShake.Server.IntegrationTests", "Tests\GraphQLStrawberryShake.Server.IntegrationTests\GraphQLStrawberryShake.Server.IntegrationTests.csproj", "{9DA8F1AA-3441-4719-9C3D-495205F745F3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQLStrawberryShake.Tests", "Tests\GraphQLStrawberryShake.Tests\GraphQLStrawberryShake.Tests.csproj", "{6FEB149A-07B6-472D-843E-8C9F098DF86D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{B69E60E4-E1F3-4AD5-81F6-2677120AD98C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,14 +23,10 @@ Global
 		{EEDE42E8-5BD9-4B16-91FE-8D68AF8B1343}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EEDE42E8-5BD9-4B16-91FE-8D68AF8B1343}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EEDE42E8-5BD9-4B16-91FE-8D68AF8B1343}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9DA8F1AA-3441-4719-9C3D-495205F745F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9DA8F1AA-3441-4719-9C3D-495205F745F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9DA8F1AA-3441-4719-9C3D-495205F745F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9DA8F1AA-3441-4719-9C3D-495205F745F3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6FEB149A-07B6-472D-843E-8C9F098DF86D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6FEB149A-07B6-472D-843E-8C9F098DF86D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6FEB149A-07B6-472D-843E-8C9F098DF86D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6FEB149A-07B6-472D-843E-8C9F098DF86D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B69E60E4-E1F3-4AD5-81F6-2677120AD98C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B69E60E4-E1F3-4AD5-81F6-2677120AD98C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B69E60E4-E1F3-4AD5-81F6-2677120AD98C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B69E60E4-E1F3-4AD5-81F6-2677120AD98C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
